Move LEM-only gadgets to `lem/gadgets.rs`
